+++ processor/src/test/resources/expected/com/example/basic/React4j_GenericTypeComponent.java
@@ -52,7 +52,7 @@ abstract class React4j_GenericTypeComponent<T> extends GenericTypeComponent<T> {
   }
 
   private static final class LiteNativeReactComponent<T> extends NativeComponent implements OnShouldComponentUpdate {
-    private React4j_GenericTypeComponent $$react4j$$_component;
+    private React4j_GenericTypeComponent<T> $$react4j$$_component;
 
     @JsConstructor
     LiteNativeReactComponent(@Nullable final JsPropertyMap<Object> props) {
@@ -73,7 +73,7 @@ abstract class React4j_GenericTypeComponent<T> extends GenericTypeComponent<T> {
   }
 
   private static final class NativeReactComponent<T> extends NativeComponent implements OnShouldComponentUpdate, OnComponentWillUnmount {
-    private React4j_GenericTypeComponent $$react4j$$_component;
+    private React4j_GenericTypeComponent<T> $$react4j$$_component;
 
     @JsConstructor
     NativeReactComponent(@Nullable final JsPropertyMap<Object> props) {